Packages updated to use new CACHE variable, new clear and install scripts

+++ packages/linux-api-3.19/build.cmd.sh
@@ -1,12 +1,13 @@
 
-pushd $SRC/*/ || die
+[ "$TARGET" = "initsys" ] || die "Package $NAME doesn't support $TARGET target!"
+
+cd $SRC/*/ || die
 
 make mrproper
 
 # Reset fake root directory
 rm -rf $BUILD &> /dev/null
-mkdir -v -p $BUILD
+mkdir -v -p $BUILD/tools
 
-make INSTALL_HDR_PATH=$BUILD headers_install
+make INSTALL_HDR_PATH=$BUILD/tools headers_install
 
-popd
